Return a pointer from event.StateKey
Returning a pointer should make it harder for people using the type to
forget to check that the event is a state event. Using the empty string
as a default value is a bad idea because it is also a valid state key
for room level state events.

Add a StateKeyEquals convienience method to make the checks easier.