Make KeyFetcher interface return keys
This commit reworks the KeyFetcher interface so that, for each key requested,
it returns that key in the form of a PublicKeyLookupRequest, rather than the
raw JSON which came from the /keys/v1/query response.

This will give us a better basis on which to implement fetching keys from
perspectives servers.

The corollary is that we also need to update KeyDatabase to work at the higher
level (of keys instead of JSON).

(We'll need to update the application level to handle the changes in
interfaces).