Make KeyFetcher interface return keys (#71)
* Factor out Encode, Decode from Base64String

separate these out from the JSON marshall/unmarshall methods so that they can
be used elsewhere without the whole JSON palaver

* Add TODO

we shouldn't re-request expired keys

* Make KeyFetcher interface return keys

This commit reworks the KeyFetcher interface so that, for each key requested,
it returns that key in the form of a PublicKeyLookupRequest, rather than the
raw JSON which came from the /keys/v1/query response.

This will give us a better basis on which to implement fetching keys from
perspectives servers.

The corollary is that we also need to update KeyDatabase to work at the higher
level (of keys instead of JSON).

(We'll need to update the application level to handle the changes in
interfaces).

* Clean up response from LookupServerKeys

There's not much point in having LookupServerKeys stick its results into a map
which we then unpick and rebuild. Just have it return the results list.