breaking: Modify the API for ReverseTopologicalOrdering and LoadAndVerify
They now both require an `Ordering` which is usually `OrderingAuthEvents`
except in the case of backfilling, where we use `OrderingPrevEvents`.
We need to sort by prev_events for backfilling so we can make optimisations
and not call `/state_ids` repeatedly, as well as work out the state at
that event for snapshots/PDU checks.