Update `applyAndAuthEvents` (#336)
This updates `applyAndAuthEvents` in state res v2 to pick auth events from the event if there is no matching event type/state key tuple in the partial state. It also optimises the auth event provider so that it only unmarshals the contents of events if they have changed to save CPU/allocations.